docs: add architecture analysis and refactor domain logic
- Created `docs/refactoring/ANALYSIS_AND_PLAN.md` to map out components, URLs, data flow, and workflow for contributors.
- Separated pure geometry functions from `frontend/src/lib/utils.js` into `frontend/src/domain/geometry.js` to isolate spatial calculations.
- Re-exported geometry functions from `utils.js` to preserve backward compatibility.
- Added JSDoc `@typedef` annotations for `Asset`, `Entity`, and `Instance` in `frontend/src/domain/assetService.js` to enhance type safety.
- Refactored `DesignCanvas.jsx` to separate view concerns by extracting state and interaction logic into a new custom hook `useCanvasInteraction`.
- Verified changes with UI rendering tests and Go unit tests.
